feat: add status killed to funcrion attack
